Identify interpreters by a unique ID
When adding concurrent execution, there will be different
interpreters working together (at least one single threaded
interpreter for basic actions and a multi-threaded one for
scheduling).  To catch bugs caused by resetting the state of
an interpreter to a memento created for another interpreter we
add a unique ID to the INTERPRETER class and check that the ID
of the memento matches that of the interpreter.

* Add functions RANDOM-HEX-LIST, MAKE-UUID and
  MAKE-UUID-SYMBOL.
* Add a slot INTERPRETER-UUID to classes INTERPRETER and
  BASIC-INTERPRETER-MEMENTO.

Additionally:
* Fix some errors in the Readme